Add unified start/help/menu module
